Changing integer for double in OilChangeEntity

+++ src/main/java/com/app/usochicamochabackend/update/application/dto/ConsolidateMotorOilDTO.java
@@ -9,7 +9,7 @@ public record ConsolidateMotorOilDTO(
         Long id,
         String type,
         BrandEntity brand,
-        Integer quantity,
+        Double quantity,
         Integer averageChangeHours,
         LocalDate dateLastUpdate,
         Double hourMeterLastUpdate,
